Remove server model
