+++ bin/populate_sample_db
@@ -6,7 +6,6 @@ sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), '../'))
 
 from scoring_engine.models.team import Team
 from scoring_engine.models.user import User
-from scoring_engine.models.server import Server
 from scoring_engine.models.service import Service
 from scoring_engine.models.property import Property
 from scoring_engine.models.round import Round
@@ -27,12 +26,10 @@ services = []
 
 print("Creating models and writing them to DB")
 num_teams = 6
-num_servers = randint(2, 4)
 num_services = randint(1, 3)
 num_properties = randint(2, 4)
 num_rounds = randint(20, 50)
 print("\tNum Teams: " + str(num_teams))
-print("\tNum Servers: " + str(num_servers))
 print("\tNum Services: " + str(num_services))
 print("\tNum Properties: " + str(num_properties))
 print("\tNum Rounds: " + str(num_rounds))
@@ -49,21 +46,16 @@ for num in range(1, num_teams + 1):
         user = User(username="user" + str(user_num), password="testpass", team=team)
         db.save(user)
 
-    for server_num in range(1, num_servers + 1):
-        print("\t\tCreating Server " + str(server_num))
-        server = Server(name="Example Server " + str(server_num), team=team)
-        db.save(server)
+    for service_num in range(1, num_services + 1):
+        print("\t\t\tCreating Service " + str(service_num))
+        service = Service(name="ICMP IPv4", team=team, check_name="ICMP IPv4 Check")
+        db.save(service)
+        services.append(service)
 
-        for service_num in range(1, num_services + 1):
-            print("\t\t\tCreating Service " + str(service_num))
-            service = Service(name="ICMP IPv4", server=server, check_name="ICMP IPv4 Check")
-            db.save(service)
-            services.append(service)
-
-            for property_num in range(1, num_properties + 1):
-                print("\t\t\t\tCreating Property " + str(property_num))
-                property_obj = Property(name="prop" + str(property_num), value="127.0.0." + str(property_num), service=service)
-                db.save(property_obj)
+        for property_num in range(1, num_properties + 1):
+            print("\t\t\t\tCreating Property " + str(property_num))
+            property_obj = Property(name="prop" + str(property_num), value="127.0.0." + str(property_num), service=service)
+            db.save(property_obj)
 
 print("Creating " + str(num_rounds) + " rounds")
 for num_round in range(1, num_rounds + 1):
