+++ scoring_engine/db.py
@@ -26,7 +26,6 @@ class DB(object):
         self.destroy()
         from scoring_engine.models.user import User
         from scoring_engine.models.team import Team
-        from scoring_engine.models.server import Server
         from scoring_engine.models.service import Service
         from scoring_engine.models.check import Check
         from scoring_engine.models.property import Property
