+++ scoring_engine/models/__init__.py
@@ -1,6 +1,5 @@
 from scoring_engine.models.team import Team
 from scoring_engine.models.user import User
-from scoring_engine.models.server import Server
 from scoring_engine.models.service import Service
 from scoring_engine.models.property import Property
 from scoring_engine.models.check import Check
