+++ scoring_engine/models/service.py
@@ -9,7 +9,7 @@ class Service(Base):
     id = Column(Integer, primary_key=True)
     name = Column(String(50), nullable=False)
     check_name = Column(String(50), nullable=False)
-    server_id = Column(Integer, ForeignKey('servers.id'))
-    server = relationship("Server", back_populates="services")
+    team_id = Column(Integer, ForeignKey('teams.id'))
+    team = relationship("Team", back_populates="services")
     properties = relationship("Property", back_populates="service")
     checks = relationship("Check", back_populates="service")
