+++ scoring_engine/models/team.py
@@ -9,5 +9,5 @@ class Team(Base):
     id = Column(Integer, primary_key=True)
     name = Column(String(20), nullable=False)
     color = Column(String(10), nullable=False)
-    servers = relationship("Server", back_populates="team")
+    services = relationship("Service", back_populates="team")
     users = relationship("User", back_populates="team")
